backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyMuPDF
qdrant-client
python-dotenv
tqdm
requests
google-genai
